phase-2-rag/02-advanced-rag/requirements.txt
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-text-splitters>=0.2.0
pypdf>=4.0.0
python-dotenv>=1.0.0
litellm>=1.40.0
rich>=13.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
rank-bm25>=0.2.2
numpy>=1.24.0
ragas>=0.2.0
datasets>=2.14.0
langchain>=0.2.0
langchain-community>=0.2.0
